make coredatafixture shared
