+++ wire-ios/Wire-iOS Tests/AvailabilityTitleViewTests.swift
@@ -121,11 +121,3 @@ class AvailabilityTitleViewTests: ZMSnapshotTestCase {
     }
 
 }
-
-extension ZMUser {
-    func updateAvailability(_ newValue: AvailabilityKind) {
-        self.willChangeValue(forKey: AvailabilityKey)
-        self.setPrimitiveValue(NSNumber(value: newValue.rawValue), forKey: AvailabilityKey)
-        self.didChangeValue(forKey: AvailabilityKey)
-    }
-}
