+++ wire-ios/Wire-iOS Tests/MentionsHandlerTests.swift
@@ -131,15 +131,3 @@ final class MentionsHandlerTests: XCTestCase {
     }
 
 }
-
-extension NSAttributedString {
-    var allAttachments: [NSTextAttachment] {
-        var attachments = [NSTextAttachment]()
-        enumerateAttributes(in: wholeRange, options: []) { attributes, _, _ in
-            if let attachment = attributes[NSAttributedString.Key.attachment] as? NSTextAttachment {
-                attachments.append(attachment)
-            }
-        }
-        return attachments
-    }
-}
