+++ wire-ios/Wire-iOS Tests/Utils/CoreDataSnapshotTestCase.swift
@@ -18,10 +18,6 @@
 
 import XCTest
 
-struct SelfProvider: SelfUserProvider {
-    let selfUser: UserType & ZMEditableUser
-}
-
 /// This class provides a `NSManagedObjectContext` in order to test views with real data instead
 /// of mock objects.
 class CoreDataSnapshotTestCase: ZMSnapshotTestCase {
