+++ wire-ios/Wire-iOS UnitTests/AvailabilityStringBuilderTests.swift
